# spring-server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Student Management System - OpenAPI 3.0
  version: 1.0.11
servers:
- url: /api/v3
...
paths:
  /student:
    get:
      description: Returns list of students
      operationId: get_Student_lisT
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Student'
                type: array
                x-content-type: application/json
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Student not found
      summary: Get list of all students
      tags:
      - student
      x-accepts: application/json
      x-tags:
      - tag: student
      x-openapi-router-controller: swagger_server.controllers.student_controller
    post:
      description: Add a new student
      operationId: add_student
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        description: Create a new student
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                $ref: '#/components/schemas/Student'
          description: Successful operation
        "405":
          description: Invalid input
      summary: Add a new student
      tags:
      - student
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: student
      x-openapi-router-controller: swagger_server.controllers.student_controller
  /student/{studentId}:
    delete:
      description: delete a student
      operationId: delete_student
      parameters:
      - description: Student id to delete
        explode: false
        in: path
        name: studentId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: successful operation
        "400":
          description: Invalid student value
      summary: Deletes a student
      tags:
      - student
      x-accepts: application/json
      x-tags:
      - tag: student
      x-openapi-router-controller: swagger_server.controllers.student_controller
    get:
      description: Returns a single student
      operationId: get_student
      parameters:
      - description: ID of student to return
        explode: false
        in: path
        name: studentId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Student not found
      summary: Find student by ID
      tags:
      - student
      x-accepts: application/json
      x-tags:
      - tag: student
      x-openapi-router-controller: swagger_server.controllers.student_controller
    put:
      description: Update an existing student by Id
      operationId: update_student
      parameters:
      - description: Student id to delete
        explode: false
        in: path
        name: studentId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        description: Update an existent student
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Student not found
        "405":
          description: Validation exception
      summary: Update an existing student
      tags:
      - student
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: student
      x-openapi-router-controller: swagger_server.controllers.student_controller
...
components:
...
  schemas:
    Student:
      example:
        address: Pune
        phone: "1234567890"
        name: Ramesh
        id: 1
        email: ramesh@email.com
      properties:
        id:
          example: 1
          format: int64
          type: integer
        name:
          example: Ramesh
          type: string
        address:
          example: Pune
          format: int64
          type: string
        email:
          example: ramesh@email.com
          type: string
        phone:
          example: "1234567890"
          type: string
      type: object
      xml:
        name: student
...